backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
google-generativeai==0.3.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
google-auth==2.23.4
